ツイート削除機能の修正(1)

+++ app/controllers/api/v1/tweets_controller.rb
@@ -51,7 +51,12 @@ module Api
 
       def destroy
         tweet = Tweet.find(params[:id])
+        # head でステータスコードのみ返す
+        return head :not_found unless tweet
+        return head :forbidden unless tweet.user.id === current_api_v1_user.id
+
         tweet.destroy
+        head :no_content
       end
 
       private
